Better refactoring of logic
